Add files via upload
Restaurant-Review-Rating-Predictions is a comprehensive project that analyzes restaurant reviews, predicts ratings based on customer comments, and offers sentiment analysis to understand the emotional context. Through insightful pie charts and personalized prompts for each restaurant, Restaurant-Review-Rating-Predictions provides a holistic exploration of the culinary world, making it an indispensable tool for food enthusiasts, restaurant owners, and anyone interested in the predictive power of data analysis in the realm of dining experiences.